test: hoist patientScopedToken into shared mcpTestHelpers
