+++ src/tests/mcp/mcpEndpoint.integration.test.js
@@ -37,33 +37,12 @@ const {
     idsInBundle,
     minimalSecurity,
     personUuid,
+    patientScopedToken,
     makePatient,
     makeLocation,
     makeObservation,
     makePerson
 } = require('./mcpTestHelpers');
-
-/**
- * A patient-scoped JWT payload whose clientFhirPersonId/bwellFhirPersonId is the given
- * personId's Person._uuid (SEC-1580 IDG-5, #2481, tightened this to an exact _uuid match with
- * no _sourceId fallback -- see personUuid's doc comment in mcpTestHelpers.js). Full user/access
- * grants are included alongside the patient scope so that scope *narrowing* (not an outright
- * access denial) is what's under test -- mirrors
- * src/tests/graphqlv2/observation/observation.test.js's getGraphQLHeadersWithPerson usage and
- * src/tests/patientScope/search_with_clientfhirpersonid's jwt_payload shape.
- */
-function patientScopedToken (personId, overrides = {}) {
-    return getHeadersWithCustomPayload({
-        scope: 'patient/*.read user/*.* access/*.*',
-        username: `${personId}@example.com`,
-        clientFhirPersonId: personUuid(personId),
-        clientFhirPatientId: 'clientFhirPatient',
-        bwellFhirPersonId: personUuid(personId),
-        bwellFhirPatientId: 'bwellFhirPatient',
-        token_use: 'access',
-        ...overrides
-    }).Authorization.replace(/^Bearer /, '');
-}
 
 /**
  * createTestRequest's fnUpdateContainer callback only has an effect on the *first* call to
